osa4 updated to 4.23

+++ osa4/blogilista/controllers/blogs.js
@@ -1,16 +1,19 @@
 const blogsRouter = require('express').Router()
 const Blog = require('../models/blog')
 const User = require('../models/user')
+const jwt = require('jsonwebtoken')
+const { userExtractor } = require('../utils/middleware')
 
 blogsRouter.get('/', async (request, response) => {
   const blogs = await Blog.find({}).populate('user', { username: 1, name: 1, id: 1 })
   response.json(blogs)
 })
 
-blogsRouter.post('/', async (request, response) => {
+blogsRouter.post('/', userExtractor, async (request, response) => {
+  try {
   const body = request.body
-
-  const user = await User.findById(body.userId)
+    
+  const user = await User.findById(request.user) 
 
   const blog = new Blog({
     title: body.title,
@@ -20,7 +23,6 @@ blogsRouter.post('/', async (request, response) => {
     user: user._id
   })
 
-  try {
     const savedBlog = await blog.save()
     user.blogs = user.blogs.concat(savedBlog._id)
     await user.save()
@@ -31,9 +33,21 @@ blogsRouter.post('/', async (request, response) => {
   }
 })
 
-blogsRouter.delete('/:id', async (request, response) => {
-  await Blog.findByIdAndDelete(request.params.id)
-  response.status(204).end()
+blogsRouter.delete('/:id', userExtractor, async (request, response) => {
+  try {
+    const blog = await Blog.findById(request.params.id)
+
+    if (blog.user.toString() === request.user.toString()) {
+      await Blog.findByIdAndDelete(request.params.id)
+      response.status(204).end()
+      }
+    else {
+      return response.status(401).json({ error: 'wrong user' })
+    }
+  }
+  catch(error) {
+    response.status(400).json()
+  }
 })
 
 blogsRouter.put('/:id', async (request, response) => {
